Remove edx-platform from setup.py dependencies.
That was silly, edx-platform is not a PyPi package, no need to list this.

On the other hand, this depends on python-saml, which is also not on PyPi.
We can install it with pip because pip supports installing from git,
but I don't think setuptools support that. I am a bit confused about the
relationship between dependencies in setup.py and requirements.txt in libraries,
so I won't specify dependency on python-saml for now.